Add international designator/COSPAR ID and observer location/radius to search; redesign search UI; break out utils.py into separate utility files
