+++ repository/templates/repository/download-data.html
@@ -5,7 +5,7 @@
 <div class="container-fluid p-3">
     <div class="container-md">
         <h2 class="text-center p-2">Download data</h2>
-        <p>The entire SCORE dataset can be downloaded in CSV format for comprehensive access
+        <p class="p-3">The entire SCORE dataset can be downloaded in CSV format for comprehensive access
             to satellite brightness and position observation data. However, we recommend that
             users interested in specific subsets of data visit the <a href="{% url 'search' %}">
                 Search</a> page. The Search page allows for customized queries and returns refined
